+++ routes/users.js
@@ -5,8 +5,6 @@ const router = express.Router();
 const User = require("../models/User");
 const Group = require("../models/Group");
 
-const { getGroupNotice } = require("./controllers/noticController");
-
 const ERROR = require("../constants/error");
 
 router.get("/:user_id", async (req, res) => {
@@ -42,10 +40,12 @@ router.get("/:user_id/groups", async function (req, res, next) {
     }
 
     if (userInfo.role === "MEMBER") {
-      return res.status(200).send(userInfo.groups);
+      return res.status(200).json(userInfo.groups);
     }
   } catch (err) {
-    res.status(400).json({ message: ERROR.GROUP_NOT_FOUND });
+    err.status = 400;
+    err.message = ERROR.GROUP_NOT_FOUND;
+    next(err);
   }
 });
 
@@ -164,6 +164,47 @@ router.post(
   }
 );
 
-router.get("/:user_id/groupNotice", getGroupNotice);
+router.get("/:user_id/groupNotice", async function (req, res, next) {
+  try {
+    const { user_id } = req.params;
+
+    const result = await User.find({
+      _id: user_id,
+    }).populate({
+      path: "groups.groupId",
+      populate: {
+        path: "notices",
+        match: {
+          "notices.period.startDate": {
+            $lte: new Date().toLocaleDateString(),
+          },
+          "notices.period.endDate": { $gte: new Date().toLocaleDateString() },
+        },
+      },
+    });
+
+    if (result[0].groups.length > 0) {
+      const myGroupList = [];
+
+      result[0].groups.map((group) => {
+        if (group.status === "PARTICIPATING") {
+          const { name, notices, colorCode } = group.groupId;
+
+          myGroupList.push({
+            name,
+            notices: [...notices],
+            colorCode,
+          });
+        }
+      });
+
+      res.status(200).json({ myGroupList });
+    } else {
+      res.status(200).json({ name: "", notices: [], colorCode: "" });
+    }
+  } catch (err) {
+    next(err);
+  }
+});
 
 module.exports = router;
